pyproject.toml
[project]
name = "upskill-hub-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asyncpg>=0.30.0",
    "certifi==2025.1.31",
    "click==8.1.8",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi-cli==0.0.7",
    "fastapi[standard]==0.115.11",
    "greenlet>=3.1.1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "loguru>=0.7.3",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pygments==2.19.1",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rich==13.9.4",
    "rich-toolkit==0.13.2",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlalchemy>=2.0.39",
    "sqlmodel>=0.0.24",
    "starlette==0.46.1",
    "typer==0.15.2",
    "typing-extensions==4.12.2",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchfiles==1.0.4",
    "websockets==15.0.1",
]
